pyproject.toml
[tool.poetry]
name = "pe-org-air-platform"
version = "1.0.0"
description = "PE Org-AI-R Platform - AI Readiness Assessment for Private Equity"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
snowflake-connector-python = "^3.6.0"
redis = "^5.0.1"
boto3 = "^1.34.0"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
# CS2 new dependencies
sec-edgar-downloader = "^5.0.0"
pdfplumber = "^0.10.3"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"
structlog = "^24.1.0"
python-jobspy = "^1.0"
streamlit = "^1.33.0"
scrapfly-sdk = "^0.8"
alembic = "^1.18.4"
snowflake-sqlalchemy = "^1.8.2"
weasyprint = "^68.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
black = "^24.1.0"
ruff = "^0.1.14"
mypy = "^1.8.0"
fakeredis = "^2.21.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"